🐛 🐛  Fixing my copy constructor and operator = assignment
I created a tests for my operator= and found out it does not work correctly. After created a test for copy constructor I found out that the problem is in the copy constructor. Previously I did not set the parent pointers for new left and right children. Same thing had to be done in operator plus I needed to change the return type to reference instead of pointer. Now my tests correctly pass