:sparkles: :sparkles: Creating a remove by key method
Creating the remove by key method, which would have same logic as the remove we already have, however we can just use that method to avoid redundancy. The way it works it just finds the node we need to remove with pointer T class. Then we just call remove by T& item.